Enable wheels for 32/64 architecture (#184)
* [ADD] wheels

* Correct overflow for np.int

* Temporary add wheels for testing

* Missed a long!

* Docs to wheels.yaml

* Improve doc and remove redundant code

+++ ConfigSpace/hyperparameters.pxd
@@ -33,7 +33,7 @@ cdef class NumericalHyperparameter(Hyperparameter):
 
 cdef class IntegerHyperparameter(NumericalHyperparameter):
     cdef ufhp
-    cpdef long _transform_scalar(self, double scalar)
+    cpdef long long _transform_scalar(self, double scalar)
     cpdef np.ndarray _transform_vector(self, np.ndarray vector)
 
 cdef class FloatHyperparameter(NumericalHyperparameter):
